requirements.txt
pyFluidSynth==1.2.4
-e hg+http://bitbucket.org/pygame/pygame